chill_requirements.txt
backports.tarfile==1.2.0
flask==3.1.0
importlib-metadata==8.0.0
jaraco.collections==5.1.0
munch==4.0.0
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-nccl-cu12==2.21.5
nvidia-nvtx-cu12==12.4.127
phonemizer==3.3.0
pip-chill==1.0.3
platformdirs==4.2.2
scipy==1.15.1
soundfile==0.13.0
tomli==2.0.1
torchaudio==2.6.0.dev20250111+cpu
torchvision==0.22.0.dev20250111+cpu
transformers==4.48.0
triton==3.1.0
